// apps/frontend/src/shared/api/generated/model/createTodoDTO.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Clarte API Gateway
 * Gateway for microservices
 * OpenAPI spec version: 1.0
 */

export interface CreateTodoDTO {
  /** Название задачи */
  title: string;
  /** Описание задачи */
  description: string;
  /** Дата выполнения задачи (ISO-8601) */
  dueDate: string;
}